test_radar_data_queue_concurrence takes care of threaded radar_data_queue tests.
